+++ Movies.Server/ApiStartup.cs
@@ -1,4 +1,5 @@
 ﻿using GraphiQl;
+using GraphQL.Server.Ui.Playground;
 using GraphQL.Types;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;