+++ changeLog.md
@@ -12,3 +12,10 @@ All notable changes to this project will be documented in this file.
 -Movies.GrainClients.Test project
 -Movies.Grains.Test project
 -Movies.Server.Test project
+
+## [1.0.0] - 13-02-2022
+### Added
+-GrapphQL and Controller unit tests were defined into Movies.Server.Test project but not filled yet
+-Task<IEnumerable<MovieModel>> GetList(string genre); interface was added into IMovieGrainClient interface
+-Task<IEnumerable<MovieModel>> GetRatedFilms(); interface was added into IMovieGrainClient interface
+-Task Update(string key, string name); interface was added into IMovieGrainClient interface